scripts/requirements.txt
numpy
opencv-python
scikit-learn
pillow
imageio
